requirements.txt
numpy==1.26.3
opencv_python==4.9.0.80
Pillow==10.2.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
torch==2.1.2
torchvision==0.16.2
